requirements.txt
aiogram
python-dotenv
loguru
pytest
sqlalchemy
